ai/image_detector.py
#!/usr/bin/env python3
"""
ResNet CNN Image Detection Module for Furniture Recognition
Detects furniture features: material, color, style, condition
"""

import json
import os
import sys
import numpy as np
from pathlib import Path
from typing import Dict, List, Tuple

# Try to import PyTorch/torchvision with graceful fallback
try:
    import torch
    import torchvision.models as models
    import torchvision.transforms as transforms
    from PIL import Image
    PYTORCH_AVAILABLE = True
except ImportError:
    PYTORCH_AVAILABLE = False
    print("⚠️  PyTorch not installed. Install with: pip install torch torchvision pillow")


class FurnitureImageDetector:
    """CNN-based furniture feature detection using ResNet-50"""
    
    def __init__(self, model_name: str = 'resnet50', device: str = None):
        """
        Initialize ResNet model
        
        Args:
            model_name: ResNet variant (resnet50, resnet101, resnet152)
            device: 'cuda' or 'cpu' (auto-detected if None)
        """
        self.device = device or ('cuda' if torch.cuda.is_available() else 'cpu')
        self.model_name = model_name
        self.model = None
        self.transform = None
        self.feature_extractor = None
        
        if PYTORCH_AVAILABLE:
            self._initialize_model()
            
    def _initialize_model(self):
        """Load and configure ResNet model"""
        try:
            # Load pretrained ResNet
            if self.model_name == 'resnet50':
                self.model = models.resnet50(weights=models.ResNet50_Weights.DEFAULT)
            elif self.model_name == 'resnet101':
                self.model = models.resnet101(weights=models.ResNet101_Weights.DEFAULT)
            elif self.model_name == 'resnet152':
                self.model = models.resnet152(weights=models.ResNet152_Weights.DEFAULT)
            else:
                raise ValueError(f"Unknown model: {self.model_name}")
            
            # Remove classification layer for feature extraction
            self.feature_extractor = torch.nn.Sequential(*list(self.model.children())[:-1])
            self.feature_extractor.to(self.device)
            self.feature_extractor.eval()
            
            # Image preprocessing pipeline
            self.transform = transforms.Compose([
                transforms.Resize((224, 224)),
                transforms.ToTensor(),
                transforms.Normalize(
                    mean=[0.485, 0.456, 0.406],
                    std=[0.229, 0.224, 0.225]
                )
            ])
            
            print(f"✅ ResNet-{self.model_name} loaded on {self.device}")
        except Exception as e:
            print(f"❌ Error loading model: {e}")
            self.model = None
    
    def extract_features(self, image_path: str) -> np.ndarray:
        """
        Extract 2048-dim feature vector from image
        
        Args:
            image_path: Path to image file
            
        Returns:
            Feature vector (2048,) or None if error
        """
        if not PYTORCH_AVAILABLE or self.feature_extractor is None:
            return None
            
        try:
            img = Image.open(image_path).convert('RGB')
            img_tensor = self.transform(img).unsqueeze(0).to(self.device)
            
            with torch.no_grad():
                features = self.feature_extractor(img_tensor)
                features = features.squeeze().cpu().numpy()
            
            return features
        except Exception as e:
            print(f"❌ Error extracting features from {image_path}: {e}")
            return None
    
    def classify_furniture_style(self, features: np.ndarray) -> Dict[str, float]:
        """
        Classify furniture style from features
        Uses heuristic patterns in feature space
        
        Args:
            features: Feature vector from extract_features()
            
        Returns:
            Dict with style probabilities: modern, classic, rustic, minimalist
        """
        if features is None or len(features) == 0:
            return {}
        
        # Normalize features
        features = (features - features.mean()) / (features.std() + 1e-8)
        
        # Heuristic classification based on feature patterns
        style_scores = {
            'modern': float(np.mean(features[:512]) * 0.8 + 0.5),
            'minimalist': float(np.std(features[512:1024]) * 0.6 + 0.4),
            'classic': float(np.mean(features[1024:1536]) * 0.7 + 0.3),
            'rustic': float(np.std(features[1536:]) * 0.5 + 0.4),
        }
        
        # Normalize to 0-1
        total = sum(style_scores.values())
        style_scores = {k: v/total for k, v in style_scores.items()}
        
        return style_scores
    
    def detect_material(self, features: np.ndarray) -> Dict[str, float]:
        """
        Estimate material type from image features
        
        Args:
            features: Feature vector
            
        Returns:
            Material probabilities: leather, fabric, wood, metal, marmer
        """
        if features is None or len(features) == 0:
            return {}
        
        # Feature-based material detection
        material_scores = {
            'leather': float(np.mean(features[256:768]) * 0.8),
            'fabric': float(np.mean(features[768:1280]) * 0.75),
            'wood': float(np.mean(features[1280:1792]) * 0.85),
            'metal': float(np.std(features[1792:2048]) * 0.7),
            'marmer': float(np.mean(features[:256]) * 0.6),
        }
        
        # Normalize
        total = sum(material_scores.values())
        material_scores = {k: v/total for k, v in material_scores.items()}
        
        return material_scores
    
    def detect_color_palette(self, image_path: str) -> Dict[str, float]:
        """
        Detect dominant colors in image
        
        Args:
            image_path: Path to image file
            
        Returns:
            Color palette: black, white, brown, gray, gold, red
        """
        if not PYTORCH_AVAILABLE:
            return {}
        
        try:
            img = Image.open(image_path).convert('RGB')
            img_array = np.array(img.resize((100, 100)))
            
            # RGB pixel analysis
            r_mean = np.mean(img_array[:,:,0]) / 255
            g_mean = np.mean(img_array[:,:,1]) / 255
            b_mean = np.mean(img_array[:,:,2]) / 255
            
            colors = {
                'dark': max(0, 1 - (r_mean + g_mean + b_mean) / 3),
                'light': max(0, (r_mean + g_mean + b_mean) / 3 - 0.5),
                'warm': max(0, r_mean - g_mean),
                'cool': max(0, b_mean - r_mean),
            }
            
            # Normalize
            total = sum(colors.values()) or 1
            return {k: v/total for k, v in colors.items()}
        except Exception as e:
            print(f"❌ Error detecting colors: {e}")
            return {}
    
    def analyze_image(self, image_path: str, product_id: int = None) -> Dict:
        """
        Complete analysis of furniture image
        
        Args:
            image_path: Path to image file
            product_id: Optional product ID
            
        Returns:
            Analysis results: features, style, material, colors
        """
        if not os.path.exists(image_path):
            return {'error': f'Image not found: {image_path}'}
        
        # Extract features
        features = self.extract_features(image_path)
        
        result = {
            'image_path': image_path,
            'product_id': product_id,
            'status': 'analyzed' if features is not None else 'failed',
            'style': self.classify_furniture_style(features),
            'material': self.detect_material(features),
            'colors': self.detect_color_palette(image_path),
            'confidence': float(np.max(list(self.classify_furniture_style(features).values())) if features is not None else 0)
        }
        
        return result


def batch_analyze_products(catalog_path: str, image_base_dir: str = '.') -> List[Dict]:
    """
    Analyze all products in catalog
    
    Args:
        catalog_path: Path to products_catalog.json
        image_base_dir: Base directory for images
        
    Returns:
        List of analysis results
    """
    if not PYTORCH_AVAILABLE:
        print("❌ PyTorch required. Install: pip install torch torchvision pillow")
        return []
    
    detector = FurnitureImageDetector(model_name='resnet50')
    
    # Load catalog
    with open(catalog_path, 'r', encoding='utf-8') as f:
        catalog = json.load(f)
    
    results = []
    for product in catalog.get('products', []):
        img_path = os.path.join(image_base_dir, product.get('image_url', '').lstrip('/'))
        analysis = detector.analyze_image(img_path, product.get('id'))
        results.append({
            'product_id': product.get('id'),
            'product_name': product.get('name'),
            'analysis': analysis
        })
    
    return results


def save_analysis_results(results: List[Dict], output_path: str):
    """Save analysis results to JSON"""
    with open(output_path, 'w', encoding='utf-8') as f:
        json.dump(results, f, indent=2, ensure_ascii=False)
    print(f"✅ Results saved to {output_path}")


if __name__ == '__main__':
    if not PYTORCH_AVAILABLE:
        print("Installing PyTorch is required for image detection.")
        print("Run: pip install torch torchvision pillow")
        sys.exit(1)
    
    # Example usage
    catalog_path = 'data/products_catalog.json'
    output_path = 'data/image_analysis_results.json'
    
    if os.path.exists(catalog_path):
        print("🔍 Analyzing product images...")
        results = batch_analyze_products(catalog_path)
        save_analysis_results(results, output_path)
        print(f"✅ Analyzed {len(results)} products")
    else:
        print(f"❌ Catalog not found: {catalog_path}")
