ai/requirements.txt
## AI & ML Dependencies
torch==2.1.0
torchvision==0.16.0
pillow==10.0.0

## LLM APIs
google-generativeai>=0.3.0
openai>=1.3.0
anthropic>=0.7.0

## Development Tools
python-dotenv>=1.0.0
httpx>=0.24.0
aiohttp>=3.9.0

## Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

## Async/Concurrency
asyncio-contextmanager>=1.0.0
aiofiles>=23.0.0

## API Client
requests>=2.31.0
httpx>=0.24.0

## Utilities
pydantic>=2.0.0
python-json-logger>=2.0.0
